# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BanglaConnect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /connections:
    get:
      operationId: listConnections
      tags: [connections]
      summary: List current user's connections
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Connection"
    post:
      operationId: requestConnection
      tags: [connections]
      summary: Send a connection request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionInput"
      responses:
        "201":
          description: Connection request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
  /connections/{id}:
    patch:
      operationId: updateConnection
      tags: [connections]
      summary: Accept or decline a connection request
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionUpdate"
      responses:
        "200":
          description: Updated connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
...
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List news items
      parameters:
        - name: tag
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of news items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsItem"
    post:
      operationId: createNews
      tags: [news]
      summary: Submit a news item
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsInput"
      responses:
        "201":
          description: Created news item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
  /news/{id}:
    get:
      operationId: getNewsItem
      tags: [news]
      summary: Get a news item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: News item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
        "404":
          description: Not found

  # ── Jobs ───────────────────────────────────────────────────────────────────
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List job postings
      parameters:
        - name: workMode
          in: query
          required: false
          schema:
            type: string
        - name: tag
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: location
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Post a job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job posting by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job posting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found

  # ── Events ─────────────────────────────────────────────────────────────────
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List community events
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: upcoming
          in: query
          required: false
          schema:
            type: boolean
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a community event
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
  /events/{id}/rsvp:
    post:
      operationId: rsvpEvent
      tags: [events]
      summary: RSVP to an event
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: RSVP created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventRsvpStatus"
    delete:
      operationId: cancelRsvp
      tags: [events]
      summary: Cancel RSVP to an event
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: RSVP cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventRsvpStatus"

  # ── Vault ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Profile:
      type: object
      required: [id, clerkId, role, fullName, country, mentorStatus, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [student, mentor, professional, admin]
        bio:
          type: ["string", "null"]
        country:
          type: string
        city:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        professionalField:
          type: ["string", "null"]
        profession:
          type: ["string", "null"]
        specialties:
          type: array
          items:
            type: string
        mentorAvailable:
          type: boolean
        mentorStatus:
          type: string
          enum: [pending, approved, rejected]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Connection:
      type: object
      required: [id, fromProfileId, toProfileId, status, createdAt]
      properties:
        id:
          type: integer
        fromProfileId:
          type: integer
        toProfileId:
          type: integer
        status:
          type: string
          enum: [pending, accepted, declined]
        message:
          type: ["string", "null"]
        fromProfile:
          $ref: "#/components/schemas/Profile"
        toProfile:
          $ref: "#/components/schemas/Profile"
        createdAt:
          type: string
    ConnectionInput:
      type: object
      required: [toProfileId]
      properties:
        toProfileId:
          type: integer
        message:
          type: string
    ConnectionUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, declined]
...
    NewsItem:
      type: object
      required: [id, title, summary, tags, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
        content:
          type: ["string", "null"]
        externalUrl:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        authorId:
          type: ["integer", "null"]
        author:
          $ref: "#/components/schemas/Profile"
        createdAt:
          type: string
    NewsInput:
      type: object
      required: [title, summary]
      properties:
        title:
          type: string
        summary:
          type: string
        content:
          type: string
        externalUrl:
          type: string
        coverImageUrl:
          type: string
        source:
          type: string
        tags:
          type: array
          items:
            type: string

    # ── Job ──────────────────────────────────────────────────────────────────
    Job:
      type: object
      required: [id, title, company, location, jobType, workMode, description, tags, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        location:
          type: string
        jobType:
          type: string
        workMode:
          type: string
        description:
          type: string
        applyUrl:
          type: ["string", "null"]
        salary:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        postedById:
          type: ["integer", "null"]
        postedBy:
          $ref: "#/components/schemas/Profile"
        createdAt:
          type: string
        expiresAt:
          type: ["string", "null"]
    JobInput:
      type: object
      required: [title, company, location, description]
      properties:
        title:
          type: string
        company:
          type: string
        location:
          type: string
        jobType:
          type: string
        workMode:
          type: string
        description:
          type: string
        applyUrl:
          type: string
        salary:
          type: string
        tags:
          type: array
          items:
            type: string
        expiresAt:
          type: string

    # ── Event ────────────────────────────────────────────────────────────────
    Event:
      type: object
      required: [id, title, description, category, eventDate, isVirtual, isLive, tags, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        eventDate:
          type: string
        endDate:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        isVirtual:
          type: boolean
        isLive:
          type: boolean
        streamUrl:
          type: ["string", "null"]
        registrationUrl:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        organizerId:
          type: ["integer", "null"]
        organizer:
          $ref: "#/components/schemas/Profile"
        rsvpCount:
          type: integer
        userRsvped:
          type: boolean
        createdAt:
          type: string
    EventInput:
      type: object
      required: [title, description, eventDate]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        eventDate:
          type: string
        endDate:
          type: string
        location:
          type: string
        isVirtual:
          type: boolean
        isLive:
          type: boolean
        streamUrl:
          type: string
        registrationUrl:
          type: string
        coverImageUrl:
          type: string
        tags:
          type: array
          items:
            type: string
    EventRsvpStatus:
      type: object
      required: [rsvped, rsvpCount]
      properties:
        rsvped:
          type: boolean
        rsvpCount:
          type: integer

    # ── Vault ─────────────────────────────────────────────────────────────────
...