# lib/api-zod/src/generated/types/connection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BanglaConnect API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConnectionStatus } from './connectionStatus';
import type { Profile } from './profile';

export interface Connection {
  id: number;
  fromProfileId: number;
  toProfileId: number;
  status: ConnectionStatus;
  /** @nullable */
  message?: string | null;
  fromProfile?: Profile;
  toProfile?: Profile;
  createdAt: string;
}